change to postgres 'pg' npm library
