+++ micros/people/get-all-people/logic.ts
@@ -1,7 +1,13 @@
 import { getPeople } from './model';
 import { People } from './types';
+import { endPool } from '../../../libs/database';
+import { log } from '../../../libs/helpers/log';
 
 export const getPeoples = async (): Promise<People[]> => {
     const data: People[] = await getPeople();
+    log('getPeoples.data', data);
+
+    await endPool();
+
     return data;
 };
